feat: make preferences and about window transient
Currently, the preferences and the about window are
on equal level to the main window, and thus do not block
input on the main window, are not centred on it, and are
in general, not logically bound to it.

Making them transient for the main window changes this
behaviour and introduces everything above.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>