feat(ui): ellipsize in preference's combo boxes
The long names of BahnCards and search providers can
result in too large minimal widths for their combo boxes,
and thus the preferences window. As such, ellipsizing
them makes this window fit small screens.

With ellipses, it is crucial to have the most important
part of the ellipsized string to be not removed. For that,
it makes sense to move the abbreviations in the search
provider names to the beginning.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>