feat(): add more credits
Tobias Bernard shaped Railway's interface by creating
mock-ups, finding a lot of paper cuts, and being around
to discuss design ideas.

Sydney Sharpe influenced the current state of the English
translation of Railway by finding more precise wording
and discussing the interface in that regard with me.

As a co-maintainer with considerable code contributions,
I could be considered an author of Railway.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>