a11y(): improve screen reader support through custom accessibility labels and descriptions
The bookmark lists both get an accessibility label.

This adds labels to trips in search based on most of
its properties, and a description based on its
indicators for delays etc.

In the details view, it also adds similarly descriptive
labels to each trip segement, as well as their stopovers
and announcements.

Co-authored-by: Julian Schmidhuber <gitlab@schmiddi.anonaddy.com>
Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>