prof(journey-details): Improve performance with loading journey details
This was noticable when e.g. selecting a bookmarked journey or selecting
a journey from the journey list with lots of legs and stopovers.

Instead of creating all leg items in one blocking call, this now adds
them one at a time and yields between then. This happens in a
low-priority future.

This not only makes it feel faster, sysprof also indicates the
improvement, see this table plotting the largest blocking tasks before
and after (with one test each, but this already shows the improvement;
the test had 8 legs and about 50 stopovers):

| Before (ms) | After (ms) |
|-------------|------------|
|         614 |        618 |
|         104 |         39 |
|          72 |         36 |
|          71 |         35 |
|         ... |        ... |

(Note the topmost entry is shutdown, saving the bookmark)

To test this change, one can easily comment/uncomment the following line
and compare the performance when loading a bookmarked journey with lots
of legs and stopovers:

```rust
tokio::task::yield_now().await;
```