fix(journeys): Fix edge-case with highlighted journey
I was missing one edge-case when implementing <https://gitlab.com/schmiddi-on-mobile/railway/-/merge_requests/120>,
if one opened a journey (e.g. using a bookmark) and afterwards searched
for that journey such that it is included in the results, the selected
journey should be highlighted in the middle column.

This was previously not the case, this commit will fix this.