fix(a11y): remove title not supposed to be user-facing anyways
This string was just there to have the Adw.NavigationPage
have a title, because it wants one. A warning is the only
result. It is not required for anything here.

It is problematic, however, to have it, because then it
is part of the accessibility tree, and hence read via
orca.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>