feat(): make first-time users aware of automatic bookmark deletion with a toast
It makes sense to not be enabled by default so that users
are not negatively surprised by it. On the other hand, it
would make sense for most use cases of bookmarks.

That is why, this adds a toast that is shown only once
(based on a schema setting), that nudges user to turn
said setting on.

Closes: https://gitlab.com/schmiddi-on-mobile/railway/-/issues/33

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>