fix(): display first day header when initial day deviates from request
When the search results display trips for the requested
day, the day header is rightly hidden. Also, when
scrolling up, it is helpful not to be shown, as then it
implies scrolling up will still yield new results from
the same day.

However, when the results contain no possible trip for on
the requested day, it is confusing that the UI does not
indicate the day mismatch between request and response.

To fix that, the first day header is now only hidden when
the user scrolled up, or the initial results are for the
requested day.

Closes: https://gitlab.com/schmiddi-on-mobile/railway/-/issues/84

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>