chores(blueprint): Port to blueprint-compiler
This ports the project from the old UI files to blueprint-compiler,
which was a much nicer syntax (but which still has minor things
missing). Most of the job was done by the porting-tool
blueprint-compiler has, but expressions and breakpoints are not yet
supported by that tool and had to be done manually.

Note that currently expressions as properties is not supported by
blueprint-compiler, I therefore had to move the discount card dropdown
in the search options window into Rust.

Porting the code also revealed one blueprint-compiler error, which was
reported and fixed https://gitlab.gnome.org/jwestman/blueprint-compiler/-/merge_requests/197,
and a missing feature to allow duplicate accessibility properties,
which was fixed in https://gitlab.gnome.org/jwestman/blueprint-compiler/-/merge_requests/205.
Both were released in blueprint-compiler version 0.14.0, but is not
published in most distributions yet.

See also https://jwestman.pages.gitlab.gnome.org/blueprint-compiler/.

Co-authored-by: Markus Göllnitz <camelcasenick@bewares.it>