chores(): Port backend to the properties macro
Currently, we implement properties of objects mostly "by hand". Since
quite a while now, gtk-rs has a derive macro that can generate quite a
bit of the code. This commit ports from the "by hand" implementation to
the macro throughout Railway. Furthermore, it replaces `property(...)`
invokations with the type-safe generated functions from the macro.

This is still work-in-progress. Currently, only the backend has been
ported. The frontend and all `property(...)` invokations still require
some more work. Furthermore, in many places where the properties macros
currently have type `Option<...>`, the `Option` may be removed. This is
not yet done, this is just a "stupid" translation from the old to the new
code.