fix(background-timer): terminate background loop after arrival
Without this, if the background timer ran on a journey
that was already in the past, the timer duration was
determined to be 0 resulting in an infinite loop taking
potentially a whole CPU thread.

Annoyingly, this timer is triggered before the main
window opens, if a journey was kept in the journey store
which is over, e.g. with delete-old = False or if the
journey ended shortly, within the deletion-time period.

Thus, this can even lead to Railway never starting.

Fixes: 42ee5e57102ad791f372ecf5b2662175a43fa920 ("feat(live-update-box): Add box displaying journey status and add notifications")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>