fix(journey-details-page): let journeys signal updates and use that to reload the details page
Journey reloading used to live in the details page
instead of in the journeys themselves. Moving it to the
more sensible location being the journeys, however,
led to the details page not setting the new journey, but
instead the journey being updated.

This broke with the details page's assumption that a
journey update corresponds with updating the journey
property.

Because journeys are not immutable, introducing a new
signal to them signalling an update is a proper solution
for this issue.

Fixes: 42ee5e57102ad791f372ecf5b2662175a43fa920 ("feat(live-update-box): Add box displaying journey status and add notifications")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>